pyproject.toml
[project]
name = "aitestkit"
version = "1.0.0"
description = "AI-Powered Multi-Framework Test Generation Toolkit"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Sefa Ertunc", email = "sefaertnc@gmail.com"}
]
keywords = [
    "testing", "ai", "automation", "claude", "qa",
    "test-generation", "pytest", "playwright", "locust",
    "security-testing", "performance-testing", "bdd"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    # Core
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Template rendering
    "jinja2>=3.1.0",

    # Retry logic (for API calls)
    "tenacity>=8.2.0",

    # Memvid (semantic memory) - Phase 4
    # "memvid>=0.1.0",  # TODO: Uncomment when available on PyPI
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.0.0",

    # Code quality
    "ruff>=0.1.0",
    "black>=24.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.6.0",

    # Type stubs
    "types-PyYAML>=6.0.0",
    "types-click>=7.1.0",
]

# Framework validators (optional)
validators = [
    "robotframework>=7.0",
    "gherkin-official>=29.0.0",
]

# Sample app dependencies
sample-app = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
]

# All optional dependencies
all = [
    "aitestkit[dev]",
    "aitestkit[validators]",
    "aitestkit[sample-app]",
]

[project.scripts]
aitestkit = "aitestkit.cli:main"

[project.urls]
Homepage = "https://github.com/sefaertunc/aitestkit"
Documentation = "https://github.com/sefaertunc/aitestkit#readme"
Repository = "https://github.com/sefaertunc/aitestkit"
Issues = "https://github.com/sefaertunc/aitestkit/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aitestkit = [
    "prompts/**/*.md",
    "prompts/**/*.yaml",
    "scenario/templates/*.yaml",
]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["aitestkit"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
)/
'''

# MyPy configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = ["anthropic.*", "rich.*"]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/aitestkit"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true
